+++ opt-env-conf-example/src/OptEnvConf/Example.hs
@@ -13,6 +13,7 @@ import Data.Text (Text)
 import OptEnvConf
 import Path
 import Paths_opt_env_conf_example (version)
+import System.Exit
 
 exampleMain :: IO ()
 exampleMain = do
@@ -73,7 +74,13 @@ instance HasParser Dispatch where
                 argument,
                 metavar "STR"
               ],
-        command "delete" "Delete" $ pure DispatchDelete
+        command "delete" "Delete" $
+          -- This requireCapability and exitFailure are just here as an example, and used in a check.
+          -- [ref:CapabilityCheck]
+          -- Normally this would just be "pure DispatchDelete".
+          requireCapability "exit" $
+            mapIO (const exitFailure) $
+              pure DispatchDelete
       ]
 
 data Settings = Settings